llm-workflow-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-workflow-orchestrator"
version = "0.1.0"
description = "Lightweight LLM workflow orchestrator on Temporal with full OpenTelemetry tracing"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Terence Schumacher" }]
dependencies = [
    "temporalio[opentelemetry]>=1.5",
    "openai>=1.0",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
    "opentelemetry-instrumentation-httpx>=0.43",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
]

[project.scripts]
llm-workflow-worker = "llm_workflow_orchestrator.worker:main"
llm-workflow-start = "llm_workflow_orchestrator.starter:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
